improv: reorganize code
